re-centralize a lot of SnapshotSupport code
