+++ service/src/main/java/org/databiosphere/workspacedataservice/dataimport/snapshotsupport/RawlsSnapshotSupport.java
@@ -1,21 +1,16 @@
 package org.databiosphere.workspacedataservice.dataimport.snapshotsupport;
 
-import bio.terra.workspace.model.DataRepoSnapshotAttributes;
 import bio.terra.workspace.model.DataRepoSnapshotResource;
-import com.google.common.annotations.VisibleForTesting;
-import java.util.ArrayList;
+import bio.terra.workspace.model.ResourceAttributesUnion;
+import bio.terra.workspace.model.ResourceDescription;
+import bio.terra.workspace.model.ResourceList;
 import java.util.List;
-import java.util.Objects;
 import java.util.UUID;
 import org.databiosphere.workspacedataservice.activitylog.ActivityLogger;
 import org.databiosphere.workspacedataservice.rawls.RawlsClient;
 import org.databiosphere.workspacedataservice.rawls.SnapshotListResponse;
 import org.databiosphere.workspacedataservice.retry.RestClientRetry;
-import org.databiosphere.workspacedataservice.service.model.exception.DataImportException;
 import org.databiosphere.workspacedataservice.shared.model.WorkspaceId;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.lang.Nullable;
 
 public class RawlsSnapshotSupport extends SnapshotSupport {
 
@@ -23,8 +18,6 @@ public class RawlsSnapshotSupport extends SnapshotSupport {
   private final RawlsClient rawlsClient;
   private final RestClientRetry restClientRetry;
   private final ActivityLogger activityLogger;
-
-  private static final Logger LOGGER = LoggerFactory.getLogger(RawlsSnapshotSupport.class);
 
   public RawlsSnapshotSupport(
       WorkspaceId workspaceId,
@@ -37,6 +30,18 @@ public class RawlsSnapshotSupport extends SnapshotSupport {
     this.restClientRetry = restClientRetry;
   }
 
+  @Override
+  protected ResourceList enumerateDataRepoSnapshotReferences(int offset, int pageSize) {
+    RestClientRetry.RestCall<SnapshotListResponse> restCall =
+        (() -> rawlsClient.enumerateDataRepoSnapshotReferences(workspaceId.id(), offset, pageSize));
+    SnapshotListResponse snapshotListResponse =
+        restClientRetry.withRetryAndErrorHandling(
+            restCall, "Rawls.enumerateDataRepoSnapshotReferences");
+
+    return toResourceList(snapshotListResponse.gcpDataRepoSnapshots());
+  }
+
+  @Override
   protected void linkSnapshot(UUID snapshotId) {
     RestClientRetry.VoidRestCall voidRestCall =
         (() -> rawlsClient.createSnapshotReference(workspaceId.id(), snapshotId));
@@ -45,92 +50,36 @@ public class RawlsSnapshotSupport extends SnapshotSupport {
         user -> user.linked().snapshotReference().withUuid(snapshotId));
   }
 
-  // TODO (AJ-1705): Filter out snapshots that do NOT have purpose:policy
-  private SnapshotListResponse enumerateDataRepoSnapshotReferences(int offset, int pageSize) {
-    RestClientRetry.RestCall<SnapshotListResponse> restCall =
-        (() -> rawlsClient.enumerateDataRepoSnapshotReferences(workspaceId.id(), offset, pageSize));
-    return restClientRetry.withRetryAndErrorHandling(
-        restCall, "Rawls.enumerateDataRepoSnapshotReferences");
-  }
-
-  public List<UUID> existingPolicySnapshotIds(int pageSize) {
-    return extractSnapshotIds(listAllSnapshots(pageSize));
-  }
-
-  @VisibleForTesting
-  List<UUID> extractSnapshotIds(List<DataRepoSnapshotResource> snapshotList) {
-    return snapshotList.stream()
-        .map(this::safeGetSnapshotId)
-        .filter(Objects::nonNull)
-        .distinct()
-        .toList();
-  }
-
   /**
-   * Given a DataRepoSnapshotResource representing a snapshot reference, retrieve that snapshot's
-   * UUID.
+   * Transforms a List<DataRepoSnapshotResource> to a ResourceList. RawlsSnapshotSupport and
+   * WsmSnapshotSupport return slightly different models from their API calls; performing this
+   * translation allows the two SnapshotSupport implementations to share code elsewhere.
    *
-   * @param snapshotResource the object in which to find a snapshotId
-   * @return the snapshotId if found, else null
+   * @param dataRepoSnapshotResourceList the model returned by Rawls APIs
+   * @return the model returned by WSM APIs
    */
-  @Nullable
-  @VisibleForTesting
-  protected UUID safeGetSnapshotId(DataRepoSnapshotResource snapshotResource) {
-    DataRepoSnapshotAttributes dataRepoSnapshot = snapshotResource.getAttributes();
-    if (dataRepoSnapshot != null) {
-      String snapshotIdStr = dataRepoSnapshot.getSnapshot();
-      try {
-        return UUID.fromString(snapshotIdStr);
-      } catch (Exception e) {
+  private ResourceList toResourceList(List<DataRepoSnapshotResource> dataRepoSnapshotResourceList) {
+    // transform the input List<DataRepoSnapshotResource> to List<ResourceDescription>
+    List<ResourceDescription> resources =
+        dataRepoSnapshotResourceList.stream()
+            .map(
+                dataRepoSnapshotResource -> {
+                  ResourceAttributesUnion resourceAttributesUnion = new ResourceAttributesUnion();
+                  resourceAttributesUnion.setGcpDataRepoSnapshot(
+                      dataRepoSnapshotResource.getAttributes());
 
-        String resourceId = "unknown";
-        try {
-          resourceId = snapshotResource.getMetadata().getResourceId().toString();
-        } catch (Exception inner) {
-          // something is exceptionally funky about this resource.
-          resourceId = inner.getMessage();
-        }
-        LOGGER.warn(
-            "Processed a DataRepoSnapshotResource [%s] that did not contain a valid snapshotId"
-                .formatted(resourceId));
-      }
-    }
-    return null;
-  }
+                  ResourceDescription resourceDescription = new ResourceDescription();
+                  resourceDescription.setMetadata(dataRepoSnapshotResource.getMetadata());
+                  resourceDescription.setResourceAttributes(resourceAttributesUnion);
+                  return resourceDescription;
+                })
+            .toList();
 
-  /**
-   * Query for the full list of referenced snapshots in this workspace, paginating as necessary.
-   *
-   * @param pageSize how many references to return in each paginated request
-   * @return the full list of snapshot references in this workspace
-   */
-  @VisibleForTesting
-  protected List<DataRepoSnapshotResource> listAllSnapshots(int pageSize) {
-    int offset = 0;
-    final int hardLimit = 10000; // under no circumstances return more than this many snapshots
+    // wrap the List<ResourceDescription> inside a ResourceList
+    ResourceList resourceList = new ResourceList();
+    resourceList.setResources(resources);
 
-    List<DataRepoSnapshotResource> finalList = new ArrayList<>(); // collect our results
-
-    while (offset < hardLimit) {
-      // get a page of results
-      SnapshotListResponse thisPage = enumerateDataRepoSnapshotReferences(offset, pageSize);
-
-      List<DataRepoSnapshotResource> page = thisPage.gcpDataRepoSnapshots();
-
-      // add this page of results to our collector
-      finalList.addAll(page);
-
-      if (page.size() < pageSize) {
-        // fewer results than we requested; this is the last page of results
-        return finalList;
-      } else {
-        // bump our offset and request another page of results
-        offset += pageSize;
-      }
-    }
-
-    throw new DataImportException(
-        "Exceeded hard limit of %d for number of pre-existing snapshot references"
-            .formatted(hardLimit));
+    // and return
+    return resourceList;
   }
 }
